+++ src/max/mod.rs
@@ -1,0 +1,5 @@
+//! Maximize harness values.
+
+pub use harness::MaxHarness;
+
+pub mod harness;